feat: enhance doctor registration controller with POST method and refactor naming
